Safari-609.1.11.1/LayoutTests/crypto/subtle/ecdsa-import-spki-key-expected.txt

Test importing a SPKI ECDSA key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS publicKey.toString() is '[object CryptoKey]'
PASS publicKey.type is 'public'
PASS publicKey.extractable is true
PASS publicKey.algorithm.name is 'ECDSA'
PASS publicKey.algorithm.namedCurve is 'P-256'
PASS publicKey.usages is ['verify']
PASS successfullyParsed is true

TEST COMPLETE
